[AKS] `az aks create/update`: promt during disablement about CR deletion

+++ src/azure-cli/azure/cli/command_modules/acs/azuremonitormetrics/azuremonitorprofile.py
@@ -18,7 +18,8 @@ from azure.cli.command_modules.acs.azuremonitormetrics.helper import (
 )
 from azure.cli.command_modules.acs.azuremonitormetrics.recordingrules.create import create_rules
 from azure.cli.command_modules.acs.azuremonitormetrics.recordingrules.delete import delete_rules
-from azure.cli.core.azclierror import InvalidArgumentValueError
+from azure.cli.core.azclierror import ClientError, InvalidArgumentValueError
+from knack.prompting import prompt_y_n
 
 
 # pylint: disable=line-too-long
@@ -50,6 +51,9 @@ def link_azure_monitor_profile_artifacts(
 
 # pylint: disable=line-too-long
 def unlink_azure_monitor_profile_artifacts(cmd, cluster_subscription, cluster_resource_group_name, cluster_name):
+    msg = 'This will delete all custom CRDs created by you as the custom resource `azmonitoring.coreos.com` will also be deleted. Are you sure you want to disable?'
+    if not prompt_y_n(msg, default="n"):
+        raise ClientError("Operation cancelled by user")
     # Remove DC* if prometheus is enabled
     dc_objects_list = get_dc_objects_list(cmd, cluster_subscription, cluster_resource_group_name, cluster_name)
     delete_dc_objects_if_prometheus_enabled(cmd, dc_objects_list, cluster_subscription, cluster_resource_group_name, cluster_name)
